test: migrate layout test to serial mode
