+++ bigbluebutton-tests/playwright/layouts/layouts.spec.js
@@ -1,49 +1,41 @@
 const { test } = require('@playwright/test');
-const { FocusOnPresentation } = require('./focusOnPresentation');
-const { FocusOnVideo } = require('./focusOnVideo');
 const { encodeCustomParams } = require('../customparameters/util');
 const { PARAMETER_HIDE_PRESENTATION_TOAST } = require('../core/constants');
-const { SmartLayout } = require('./smartLayout');
-const { CustomLayout } = require('./customLayout');
-const { PushLayout } = require('./pushLayout');
+const { Layouts } = require('./layouts');
 
 const hidePresentationToast = encodeCustomParams(PARAMETER_HIDE_PRESENTATION_TOAST);
 
 const CUSTOM_MEETING_ID = 'layout_management_meeting';
 
-test.describe.parallel('Layout management', () => {
-  test('Focus on presentation', async ({ browser, context, page }) => {
-    const focusOnPresentation = new FocusOnPresentation(browser, context);
-    await focusOnPresentation.initModPage(page, true, { customMeetingId: CUSTOM_MEETING_ID, customParameter: hidePresentationToast });
-    await focusOnPresentation.initUserPage(true, context, { customParameter: hidePresentationToast });
-    await focusOnPresentation.test();
+test.describe.serial('Layout management', () => {
+  const layouts = new Layouts();
+
+  test.beforeAll(async ({ browser }) => {
+    const context = await browser.newContext();
+    const page = await context.newPage();
+    await layouts.initModPage(page, true,  { customParameter: hidePresentationToast, customMeetingId: CUSTOM_MEETING_ID });
+    await layouts.initUserPage(true, context, { customParameter: hidePresentationToast });
+    await layouts.modPage.shareWebcam();
+    await layouts.userPage.shareWebcam();
   });
 
-  test('Focus on video', async ({ browser, context, page }) => {
-    const focusOnVideo = new FocusOnVideo(browser, context);
-    await focusOnVideo.initModPage(page, true, { customMeetingId: CUSTOM_MEETING_ID, customParameter: hidePresentationToast });
-    await focusOnVideo.initUserPage(true, context, { customParameter: hidePresentationToast });
-    await focusOnVideo.test();
+  test('Focus on presentation', async () => {
+    await layouts.focusOnPresentation();
   });
 
-  test('Smart layout', async ({ browser, context, page }) => {
-    const smartLayout = new SmartLayout(browser, context);
-    await smartLayout.initModPage(page, true, { customMeetingId: CUSTOM_MEETING_ID, customParameter: hidePresentationToast });
-    await smartLayout.initUserPage(true, context, { customParameter: hidePresentationToast });
-    await smartLayout.test();
+  test('Focus on video', async () => {
+    await layouts.focusOnVideo();
   });
 
-  test('Custom layout', async ({ browser, context, page }) => {
-    const customLayout = new CustomLayout(browser, context);
-    await customLayout.initModPage(page, true, { customMeetingId: CUSTOM_MEETING_ID, customParameter: hidePresentationToast });
-    await customLayout.initUserPage(true, context, { customParameter: hidePresentationToast });
-    await customLayout.test();
+  test('Smart layout', async () => {
+    await layouts.smartLayout();
   });
 
-  test('Push layout to all', async ({ browser, context, page }) => {
-    const pushLayout = new PushLayout(browser, context);
-    await pushLayout.initModPage(page, true, { customMeetingId: CUSTOM_MEETING_ID, customParameter: hidePresentationToast });
-    await pushLayout.initUserPage(true, context, { customParameter: hidePresentationToast });
-    await pushLayout.test();
+  test('Custom layout', async () => {
+    await layouts.customLayout();
+  });
+
+  test('Push layout to all', async () => {
+    await layouts.pushLayout();
   });
 });
